feat(core): allow fetchAuthSession to run with Amplify server context
